:zap: Improve AI performance
