+++ config/sentry.php
@@ -5,7 +5,7 @@ return [
 
     'environment' => env('APP_ENV', 'production'),
 
-    'release' => file_exists(base_path('VERSION')) ? env('SENTRY_PROJECT_NAME') . file_get_contents(base_path('VERSION')) : null,
+    'release' => file_exists(base_path('VERSION')) ? file_get_contents(base_path('VERSION')) : null,
 
     'traces_sample_rate' => (float) env('SENTRY_TRACES_SAMPLE_RATE', 0.2),
